+++ amfiteatr_rl/src/demo.rs
@@ -1,7 +1,7 @@
 use tch::Tensor;
 use amfiteatr_core::demo::{DemoInfoSet};
 use crate::error::TensorRepresentationError;
-use crate::tensor_data::{CtxTryIntoTensor, ConversionToTensor};
+use crate::tensor_data::{ContextTryIntoTensor, ConversionToTensor};
 
 /// Demonstration conversion to tensor meant for [`DemoInfoSet`].
 /// This is made only to demonstrate syntax for examples making sense please refer to
@@ -17,7 +17,7 @@ impl ConversionToTensor for DemoConversionToTensor {
     }
 }
 
-impl CtxTryIntoTensor<DemoConversionToTensor> for DemoInfoSet{
+impl ContextTryIntoTensor<DemoConversionToTensor> for DemoInfoSet{
     fn try_to_tensor(&self, _way: &DemoConversionToTensor) -> Result<Tensor, TensorRepresentationError> {
         Ok(Tensor::from_slice(&[1.0]))
     }